dom/tests/mochitest/general/test_interfaces.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=766694
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 766694</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=766694">Mozilla Bug 766694</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 766694 **/

// This is a list of all interfaces that are exposed to every webpage.  Please only
// add things to this list with great care.

// IMPORTANT: Do not change this list without review from a DOM peer!
var interfaceNamesInGlobalScope =
  [
    "AnimationEvent",
    "Array",
    "AsyncScrollEventDetail",
    "Attr",
    "BarProp",
    "BatteryManager",
    "BeforeUnloadEvent",
    "Blob",
    "BlobEvent",
    "CameraCapabilities",
    "CameraControl",
    "CameraManager",
    "CanvasGradient",
    "CanvasPattern",
    "CanvasRenderingContext2D",
    "CDATASection",
    "CharacterData",
    "ClientRect",
    "ClientRectList",
    "ClipboardEvent",
    "CloseEvent",
    "CommandEvent",
    "Comment",
    "CompositionEvent",
    "Controllers",
    "CRMFObject",
    "Crypto",
    "CSS2Properties",
    "CSSCharsetRule",
    "CSSConditionRule",
    "CSSFontFaceRule",
    "CSSFontFeatureValuesRule",
    "CSSGroupingRule",
    "CSSImportRule",
    "CSSMediaRule",
    "CSSMozDocumentRule",
    "CSSPageRule",
    "CSSPrimitiveValue",
    "CSSRule",
    "CSSRuleList",
    "CSSStyleDeclaration",
    "CSSStyleRule",
    "CSSStyleSheet",
    "CSSSupportsRule",
    "CSSValue",
    "CSSValueList",
    "CustomEvent",
    "DataChannel",
    "DataContainerEvent",
    "DataTransfer",
    "DesktopNotification",
    "DesktopNotificationCenter",
    "DeviceAcceleration",
    "DeviceLightEvent",
    "DeviceMotionEvent",
    "DeviceOrientationEvent",
    "DeviceProximityEvent",
    "DeviceRotationRate",
    "DeviceStorage",
    "DeviceStorageChangeEvent",
    "DeviceStorageCursor",
    "Document",
    "DocumentFragment",
    "DocumentType",
    "DOMCursor",
    "DOMError",
    "DOMException",
    "DOMImplementation",
    "DOMRequest",
    "DOMSettableTokenList",
    "DOMStringList",
    "DOMStringMap",
    "DOMTokenList",
    "DOMTransactionEvent",
    "DragEvent",
    "Element",
    "ElementReplaceEvent",
    "ElementTimeControl",
    "Event",
    "EventListenerInfo",
    "EventSource",
    "EventTarget",
    "File",
    "FileHandle",
    "FileList",
    "FileReader",
    "FileRequest",
    "FocusEvent",
    "FormData",
    "Gamepad",
    "GamepadAxisMoveEvent",
    "GamepadButtonEvent",
    "GamepadEvent",
    "GeoPositionError",
    "HashChangeEvent",
    "History",
    "HTMLAnchorElement",
    "HTMLAppletElement",
    "HTMLAreaElement",
    "HTMLAudioElement",
    "HTMLBaseElement",
    "HTMLBodyElement",
    "HTMLBRElement",
    "HTMLButtonElement",
    "HTMLCanvasElement",
    "HTMLCollection",
    "HTMLCommandElement",
    "HTMLDataListElement",
    "HTMLDirectoryElement",
    "HTMLDivElement",
    "HTMLDListElement",
    "HTMLDocument",
    "HTMLElement",
    "HTMLEmbedElement",
    "HTMLFieldSetElement",
    "HTMLFontElement",
    "HTMLFormElement",
    "HTMLFrameElement",
    "HTMLFrameSetElement",
    "HTMLHeadElement",
    "HTMLHeadingElement",
    "HTMLHRElement",
    "HTMLHtmlElement",
    "HTMLIFrameElement",
    "HTMLImageElement",
    "HTMLInputElement",
    "HTMLLabelElement",
    "HTMLLegendElement",
    "HTMLLIElement",
    "HTMLLinkElement",
    "HTMLMapElement",
    "HTMLMediaElement",
    "HTMLMenuElement",
    "HTMLMenuItemElement",
    "HTMLMetaElement",
    "HTMLMeterElement",
    "HTMLModElement",
    "HTMLObjectElement",
    "HTMLOListElement",
    "HTMLOptGroupElement",
    "HTMLOptionElement",
    "HTMLOptionsCollection",
    "HTMLOutputElement",
    "HTMLParagraphElement",
    "HTMLParamElement",
    "HTMLPreElement",
    "HTMLProgressElement",
    "HTMLPropertiesCollection",
    "HTMLQuoteElement",
    "HTMLScriptElement",
    "HTMLSelectElement",
    "HTMLSourceElement",
    "HTMLStyleElement",
    "HTMLTableCaptionElement",
    "HTMLTableCellElement",
    "HTMLTableColElement",
    "HTMLTableElement",
    "HTMLTableRowElement",
    "HTMLTableSectionElement",
    "HTMLTextAreaElement",
    "HTMLTitleElement",
    "HTMLUListElement",
    "HTMLUnknownElement",
    "HTMLVideoElement",
    "IDBCursor",
    "IDBCursorWithValue",
    "IDBDatabase",
    "IDBFactory",
    "IDBIndex",
    "IDBKeyRange",
    "IDBObjectStore",
    "IDBOpenDBRequest",
    "IDBRequest",
    "IDBTransaction",
    "IDBVersionChangeEvent",
    "ImageData",
    "JSON",
    "KeyEvent",
    "LoadStatus",
    "LocalMediaStream",
    "Location",
    "LockedFile",
    "LSProgressEvent",
    "MediaError",
    "MediaList",
    "MediaQueryList",
    "MediaStream",
    "MessageEvent",
    "MimeType",
    "MimeTypeArray",
    "ModalContentWindow",
    "MouseEvent",
    "MouseScrollEvent",
    "MozApplicationEvent",
    "MozBlobBuilder",
    "MozCanvasPrintState",
    "MozConnection",
    "MozContactChangeEvent",
    "MozCSSKeyframeRule",
    "MozCSSKeyframesRule",
    "MozMmsEvent",
    "MozMmsMessage",
    "MozMobileCellInfo",
    "MozMobileConnectionInfo",
    "MozMobileMessageManager",
    "MozMobileMessageThread",
    "MozMobileNetworkInfo",
    "MozNamedAttrMap",
    "MozNetworkStats",
    "MozNetworkStatsData",
    "MozNetworkStatsManager",
    "MozPowerManager",
    "MozSettingsEvent",
    "MozSmsEvent",
    "MozSmsFilter",
    "MozSmsMessage",
    "MozSmsSegmentInfo",
    "MozTimeManager",
    "MozTouchEvent",
    "MozWakeLock",
    "MutationEvent",
    "MutationObserver",
    "MutationRecord",
    "Navigator",
    "Node",
    "NodeFilter",
    "NodeIterator",
    "NodeList",
    "NodeSelector",
    "NotifyAudioAvailableEvent",
    "NotifyPaintEvent",
    "NSRGBAColor",
    "OfflineResourceList",
    "OpenWindowEventDetail",
    "PageTransitionEvent",
    "PaintRequest",
    "PaintRequestList",
    "PaymentRequestInfo",
    "Performance",
    "PerformanceNavigation",
    "PerformanceTiming",
    "Plugin",
    "PluginArray",
    "PopStateEvent",
    "PopupBlockedEvent",
    "ProcessingInstruction",
    "ProgressEvent",
    "PropertyNodeList",
    "Range",
    "RecordErrorEvent",
    "Rect",
    "RGBColor",
    "RTCIceCandidate",
    "RTCPeerConnection",
    "RTCSessionDescription",
    "Screen",
    "ScrollAreaEvent",
    "Selection",
    "SettingsLock",
    "SettingsManager",
    "SimpleGestureEvent",
    "SmartCardEvent",
    "SpeechRecognitionError",
    "SpeechRecognitionEvent",
    "SpeechSynthesisEvent",
    "Storage",
    "StorageEvent",
    "StorageItem",
    "StorageObsolete",
    "StyleRuleChangeEvent",
    "StyleSheet",
    "StyleSheetApplicableStateChangeEvent",
    "StyleSheetChangeEvent",
    "StyleSheetList",
    "SVGAElement",
    "SVGAltGlyphElement",
    "SVGAngle",
    "SVGAnimatedAngle",
    "SVGAnimatedBoolean",
    "SVGAnimatedEnumeration",
    "SVGAnimatedInteger",
    "SVGAnimatedLength",
    "SVGAnimatedLengthList",
    "SVGAnimatedNumber",
    "SVGAnimatedNumberList",
    "SVGAnimatedPathData",
    "SVGAnimatedPoints",
    "SVGAnimatedPreserveAspectRatio",
    "SVGAnimatedRect",
    "SVGAnimatedString",
    "SVGAnimatedTransformList",
    "SVGAnimateElement",
    "SVGAnimateMotionElement",
    "SVGAnimateTransformElement",
    "SVGAnimationElement",
    "SVGCircleElement",
    "SVGClipPathElement",
    "SVGComponentTransferFunctionElement",
    "SVGDefsElement",
    "SVGDescElement",
    "SVGDocument",
    "SVGElement",
    "SVGEllipseElement",
    "SVGEvent",
    "SVGFEBlendElement",
    "SVGFEColorMatrixElement",
    "SVGFEComponentTransferElement",
    "SVGFECompositeElement",
    "SVGFEConvolveMatrixElement",
    "SVGFEDiffuseLightingElement",
    "SVGFEDisplacementMapElement",
    "SVGFEDistantLightElement",
    "SVGFEFloodElement",
    "SVGFEFuncAElement",
    "SVGFEFuncBElement",
    "SVGFEFuncGElement",
    "SVGFEFuncRElement",
    "SVGFEGaussianBlurElement",
    "SVGFEImageElement",
    "SVGFEMergeElement",
    "SVGFEMergeNodeElement",
    "SVGFEMorphologyElement",
    "SVGFEOffsetElement",
    "SVGFEPointLightElement",
    "SVGFESpecularLightingElement",
    "SVGFESpotLightElement",
    "SVGFETileElement",
    "SVGFETurbulenceElement",
    "SVGFilterElement",
    "SVGFilterPrimitiveStandardAttributes",
    "SVGFitToViewBox",
    "SVGForeignObjectElement",
    "SVGGElement",
    "SVGGradientElement",
    "SVGImageElement",
    "SVGLength",
    "SVGLengthList",
    "SVGLinearGradientElement",
    "SVGLineElement",
    "SVGLocatable",
    "SVGMarkerElement",
    "SVGMaskElement",
    "SVGMatrix",
    "SVGMetadataElement",
    "SVGMpathElement",
    "SVGNumber",
    "SVGNumberList",
    "SVGPathElement",
    "SVGPathSeg",
    "SVGPathSegArcAbs",
    "SVGPathSegArcRel",
    "SVGPathSegClosePath",
    "SVGPathSegCurvetoCubicAbs",
    "SVGPathSegCurvetoCubicRel",
    "SVGPathSegCurvetoCubicSmoothAbs",
    "SVGPathSegCurvetoCubicSmoothRel",
    "SVGPathSegCurvetoQuadraticAbs",
    "SVGPathSegCurvetoQuadraticRel",
    "SVGPathSegCurvetoQuadraticSmoothAbs",
    "SVGPathSegCurvetoQuadraticSmoothRel",
    "SVGPathSegLinetoAbs",
    "SVGPathSegLinetoHorizontalAbs",
    "SVGPathSegLinetoHorizontalRel",
    "SVGPathSegLinetoRel",
    "SVGPathSegLinetoVerticalAbs",
    "SVGPathSegLinetoVerticalRel",
    "SVGPathSegList",
    "SVGPathSegMovetoAbs",
    "SVGPathSegMovetoRel",
    "SVGPatternElement",
    "SVGPoint",
    "SVGPointList",
    "SVGPolygonElement",
    "SVGPolylineElement",
    "SVGPreserveAspectRatio",
    "SVGRadialGradientElement",
    "SVGRect",
    "SVGRectElement",
    "SVGScriptElement",
    "SVGSetElement",
    "SVGStopElement",
    "SVGStringList",
    "SVGStylable",
    "SVGStyleElement",
    "SVGSVGElement",
    "SVGSwitchElement",
    "SVGSymbolElement",
    "SVGTests",
    "SVGTextContentElement",
    "SVGTextElement",
    "SVGTextPathElement",
    "SVGTextPositioningElement",
    "SVGTitleElement",
    "SVGTransform",
    "SVGTransformable",
    "SVGTransformList",
    "SVGTSpanElement",
    "SVGUnitTypes",
    "SVGURIReference",
    "SVGUseElement",
    "SVGViewElement",
    "SVGViewSpec",
    "SVGZoomAndPan",
    "SVGZoomEvent",
    "TCPSocket",
    "TCPServerSocket",
    "Text",
    "TextMetrics",
    "TimeEvent",
    "TimeRanges",
    "Touch",
    "TouchEvent",
    "TouchList",
    "TransitionEvent",
    "TreeWalker",
    "UIEvent",
    "UndoManager",
    "URL",
    "UserDataHandler",
    "UserProximityEvent",
    "USSDReceivedEvent",
    "ValidityState",
    "WebGLRenderingContext",
    "WebSocket",
    "WheelEvent",
    "Window",
    "WindowUtils",
    "XMLDocument",
    "XMLHttpRequest",
    "XMLHttpRequestEventTarget",
    "XMLHttpRequestUpload",
    "XPathEvaluator",
    "XPathExpression",
    "XPathNamespace",
    "XPathNSResolver",
    "XPathResult",
    "XSLTProcessor",
    "XULElement",
  ]
// IMPORTANT: Do not change this list without review from a DOM peer!

// If your interface is named nsIDOMSomeInterface and you don't mean to expose
// it to every webpage, simply change its name to nsISomeInterface to fix this problem.

  var interfaceNamesInXBLScope = [
    "BoxObject",
    "ChromeWindow",
    "DOMConstructor",
    "TreeColumn",
    "TreeColumns",
    "TreeContentView",
    "TreeSelection",
    "XULButtonElement",
    "XULCheckboxElement",
    "XULCommandDispatcher",
    "XULCommandEvent",
    "XULControlElement",
    "XULControllers",
    "XULDocument",
    "XULLabeledControlElement",
    "XULPopupElement",
    "XULTemplateBuilder",
    "XULTreeBuilder",
  ];

function runTest(win, isXBLScope) {
  for (var i in SpecialPowers.Components.interfaces) {
    var s = i.toString();
    var name = null;
    if (s.indexOf("nsIDOM") == 0) {
      name = s.substring("nsIDOM".length);
    } else if (s.indexOf("nsI") == 0) {
      name = s.substring("nsI".length);
    }
    if (name && (name in win)) {
      ok(interfaceNamesInGlobalScope.indexOf(name) >= 0 ||
         (isXBLScope && interfaceNamesInXBLScope.indexOf(name) >= 0),
         "If this is failing: DANGER, are you sure you want to expose the new interface " + name + " to all webpages as a property on the window?");
    }
  }
  for (var name of interfaceNamesInXBLScope) {
    is(name in win, isXBLScope,
       name + " should" + (isXBLScope ? "" : " NOT") + " be present in the global scope");
  }
}

runTest(window, false);
SimpleTest.waitForExplicitFinish();

</script>
<span id="span" style="-moz-binding: url(file_interfaces.xml)"></span>
</pre>
</body>
</html>
